requirements.txt
depthai-sdk
depthai-nodes
numpy